+++ tests/tests_pytorch/core/test_saving.py
@@ -55,3 +55,19 @@ def test_load_from_checkpoint_map_location_cpu_to_gpu(tmp_path, map_location):
     create_boring_checkpoint(tmp_path, BoringModel(), accelerator="cpu")
     model = BoringModel.load_from_checkpoint(f"{tmp_path}/checkpoint.ckpt", map_location=map_location)
     assert model.device.type == "cuda"
+
+
+@RunIf(min_cuda_gpus=1)
+def test_load_from_checkpoint_device_placement_with_extra_state(tmp_path):
+    """Test that the device gets chosen based on the device of the saved tensors in the checkpoint."""
+
+    class ExtraStateModel(BoringModel):
+        def get_extra_state(self):
+            return {"extra": "state"}  # state without tensors
+
+        def set_extra_state(self, state):
+            pass
+
+    create_boring_checkpoint(tmp_path, ExtraStateModel(), accelerator="cuda")
+    model = ExtraStateModel.load_from_checkpoint(f"{tmp_path}/checkpoint.ckpt", map_location=None)
+    assert model.device.type == "cuda"